requirements.txt
tensorflow==2.15.0
keras==2.15.0
matplotlib==3.7.0
numpy==1.26.4
pandas==2.1.0
scikit-learn==1.4.2
imageio==2.37.0
lazy-loader==0.4
scikit-image==0.25.2
tifffile==2025.6.11
ollama == 0.5.3
fastapi[standard]==0.116.1
accelerate
diffusers
huggingface-hub
python-dotenv
torch
transformers
transformers[sentencepiece]